+++ cishouseholds/pipeline/post_union_transformations.py
@@ -97,50 +97,11 @@ def fill_forwards(df: DataFrame) -> DataFrame:
         record_changed_value="Yes",
     )
 
-    df = fill_backwards_overriding_not_nulls(
-        df=df,
-        column_identity="participant_id",
-        ordering_column="visit_datetime",
-        dataset_column="survey_response_dataset_major_version",
-        column_list=["sex", "date_of_birth", "ethnicity"],
-    )
-    df = fill_forward_only_to_nulls(
-        df=df,
-        id="participant_id",
-        date="visit_datetime",
-        list_fill_forward=[
-            "sex",
-            "date_of_birth",
-            "ethnicity",
-        ],
-    )
     return df
 
 
 def union_dependent_cleaning(df) -> DataFrame:
     col_val_map = {
-        "ethnicity": {
-            "African": "Black,Caribbean,African-African",
-            "Caribbean": "Black,Caribbean,Afro-Caribbean",
-            "Any other Black or African or Caribbean background": "Any other Black background",
-            "Any other Black| African| Carribbean": "Any other Black background",
-            "Any other Mixed/Multiple background": "Any other Mixed background",
-            "Bangladeshi": "Asian or Asian British-Bangladeshi",
-            "Chinese": "Asian or Asian British-Chinese",
-            "English, Welsh, Scottish, Northern Irish or British": "White-British",
-            "English| Welsh| Scottish| Northern Irish or British": "White-British",
-            "Indian": "Asian or Asian British-Indian",
-            "Irish": "White-Irish",
-            "Pakistani": "Asian or Asian British-Pakistani",
-            "White and Asian": "Mixed-White & Asian",
-            "White and Black African": "Mixed-White & Black African",
-            "White and Black Caribbean": "Mixed-White & Black Caribbean",
-            "Roma": "White-Gypsy or Irish Traveller",
-            "White-Roma": "White-Gypsy or Irish Traveller",
-            "Gypsy or Irish Traveller": "White-Gypsy or Irish Traveller",
-            "Arab": "Other ethnic group-Arab",
-            "Any other white": "Any other white background",
-        },
         "participant_withdrawal_reason": {
             "Bad experience with tester / survey": "Bad experience with interviewer/survey",
             "Swab / blood process too distressing": "Swab/blood process too distressing",
@@ -248,59 +209,6 @@ def create_ever_variable_columns(df: DataFrame) -> DataFrame:
     return df
 
 
-def derive_people_in_household_count(df) -> DataFrame:
-    """
-    Correct counts of household member groups and sum to get total number of people in household. Takes maximum
-    final count by household for each record.
-    """
-    df = assign_household_participant_count(
-        df,
-        column_name_to_assign="household_participant_count",
-        household_id_column="ons_household_id",
-        participant_id_column="participant_id",
-    )
-    df = update_person_count_from_ages(
-        df,
-        column_name_to_assign="household_participants_not_consenting_count",
-        column_pattern=r"person_not_consenting_age_[1-9]",
-    )
-    df = update_person_count_from_ages(
-        df,
-        column_name_to_assign="household_members_over_2_years_and_not_present_count",
-        column_pattern=r"person_not_present_age_[1-8]",
-    )
-    df = assign_household_under_2_count(
-        df,
-        column_name_to_assign="household_members_under_2_years_count",
-        column_pattern=r"infant_age_months_[1-9]",
-        condition_column="household_members_under_2_years",
-    )
-    household_window = Window.partitionBy("ons_household_id")
-
-    household_participants = [
-        "household_participant_count",
-        "household_participants_not_consenting_count",
-        "household_members_over_2_years_and_not_present_count",
-        "household_members_under_2_years_count",
-    ]
-    for household_participant_type in household_participants:
-        df = df.withColumn(
-            household_participant_type,
-            F.max(household_participant_type).over(household_window),
-        )
-    df = df.withColumn(
-        "people_in_household_count",
-        sum_within_row(household_participants),
-    )
-    df = df.withColumn(
-        "people_in_household_count_group",
-        F.when(F.col("people_in_household_count") >= 5, "5+").otherwise(
-            F.col("people_in_household_count").cast("string")
-        ),
-    )
-    return df
-
-
 def create_formatted_datetime_string_columns(df) -> DataFrame:
     """
     Create columns with specific datetime formatting for use in output data.
@@ -401,33 +309,8 @@ def union_dependent_derivations(df) -> DataFrame:
         df = df.withColumn(
             "participant_visit_status", F.coalesce(F.col("participant_visit_status"), F.col("survey_completion_status"))
         )
-    ethnicity_map = {
-        "White": ["White-British", "White-Irish", "White-Gypsy or Irish Traveller", "Any other white background"],
-        "Asian": [
-            "Asian or Asian British-Indian",
-            "Asian or Asian British-Pakistani",
-            "Asian or Asian British-Bangladeshi",
-            "Asian or Asian British-Chinese",
-            "Any other Asian background",
-        ],
-        "Black": ["Black,Caribbean,African-African", "Black,Caribbean,Afro-Caribbean", "Any other Black background"],
-        "Mixed": [
-            "Mixed-White & Black Caribbean",
-            "Mixed-White & Black African",
-            "Mixed-White & Asian",
-            "Any other Mixed background",
-        ],
-        "Other": ["Other ethnic group-Arab", "Any other ethnic group"],
-    }
 
     df = replace_sample_barcode(df=df)
-
-    df = assign_column_from_mapped_list_key(
-        df=df, column_name_to_assign="ethnicity_group", reference_column="ethnicity", map=ethnicity_map
-    )
-    df = assign_ethnicity_white(
-        df, column_name_to_assign="ethnicity_white", ethnicity_group_column_name="ethnicity_group"
-    )
 
     df = create_ever_variable_columns(df)
